requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==24.1.0
    # via gradio
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via google-auth
aiosignal==1.4.0
    # via aiohttp
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   google-genai
    #   gradio
    #   httpx
    #   openai
    #   starlette
attrs==25.4.0
    # via
    #   aiohttp
    #   frictionless
    #   jsonschema
    #   referencing
boto3==1.42.31
    # via -r requirements.in
botocore==1.42.31
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
brotli==1.2.0
    # via gradio
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
chardet==5.2.0
    # via frictionless
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer
    #   typer-slim
    #   uvicorn
distro==1.9.0
    # via
    #   google-genai
    #   openai
docstring-parser==0.17.0
    # via google-cloud-aiplatform
et-xmlfile==2.0.0
    # via openpyxl
fastapi==0.128.0
    # via gradio
ffmpy==1.0.0
    # via gradio
filelock==3.20.3
    # via huggingface-hub
frictionless==5.18.1
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2026.1.0
    # via
    #   gradio-client
    #   huggingface-hub
google-api-core[grpc]==2.29.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
google-auth[aiohttp]==2.47.0
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   google-cloud-core
    #   google-cloud-resource-manager
    #   google-cloud-storage
    #   google-genai
google-cloud-aiplatform[all]==1.71.1
    # via
    #   -r requirements.in
    #   vertexai
google-cloud-bigquery==3.40.0
    # via google-cloud-aiplatform
google-cloud-core==2.5.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
google-cloud-resource-manager==1.16.0
    # via google-cloud-aiplatform
google-cloud-storage==2.19.0
    # via google-cloud-aiplatform
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via
    #   google-cloud-bigquery
    #   google-cloud-storage
googleapis-common-protos[grpc]==1.72.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
gradio==6.3.0
    # via -r requirements.in
gradio-client==2.0.3
    # via gradio
groovy==0.1.2
    # via gradio
grpc-google-iam-v1==0.14.3
    # via google-cloud-resource-manager
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.71.2
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   google-genai
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   safehttpx
huggingface-hub==1.3.2
    # via
    #   gradio
    #   gradio-client
humanize==4.15.0
    # via frictionless
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
isodate==0.7.2
    # via frictionless
jinja2==3.1.6
    # via
    #   frictionless
    #   gradio
jiter==0.12.0
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.26.0
    # via frictionless
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown==3.10.1
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
marko==2.2.2
    # via frictionless
markupsafe==3.0.3
    # via
    #   gradio
    #   jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
numpy==2.4.1
    # via
    #   gradio
    #   pandas
    #   shapely
openai==2.15.0
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
orjson==3.11.5
    # via gradio
packaging==25.0
    # via
    #   google-cloud-aiplatform
    #   google-cloud-bigquery
    #   gradio
    #   gradio-client
    #   huggingface-hub
pandas==2.3.3
    # via
    #   -r requirements.in
    #   gradio
petl==1.7.17
    # via frictionless
pillow==12.1.0
    # via gradio
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.27.0
    # via
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
protobuf==5.29.5
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-aiplatform
    #   google-cloud-resource-manager
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
pyasn1==0.6.2
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via
    #   fastapi
    #   frictionless
    #   google-cloud-aiplatform
    #   google-genai
    #   gradio
    #   openai
pydantic-core==2.41.5
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   frictionless
    #   google-cloud-bigquery
    #   pandas
python-multipart==0.0.21
    # via gradio
python-slugify==8.0.4
    # via frictionless
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   frictionless
    #   gradio
    #   huggingface-hub
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   frictionless
    #   google-api-core
    #   google-auth
    #   google-cloud-bigquery
    #   google-cloud-storage
    #   google-genai
rfc3986==2.0.0
    # via frictionless
rich==14.2.0
    # via typer
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
s3transfer==0.16.0
    # via boto3
safehttpx==0.1.7
    # via gradio
semantic-version==2.10.0
    # via gradio
shapely==2.1.2
    # via google-cloud-aiplatform
shellingham==1.5.4
    # via
    #   huggingface-hub
    #   typer
simpleeval==1.0.3
    # via frictionless
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   google-genai
    #   openai
starlette==0.50.0
    # via
    #   fastapi
    #   gradio
striprtf==0.0.29
    # via -r requirements.in
tabulate==0.9.0
    # via frictionless
text-unidecode==1.3
    # via python-slugify
tomlkit==0.13.3
    # via gradio
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typer==0.21.1
    # via
    #   frictionless
    #   gradio
typer-slim==0.21.1
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   fastapi
    #   frictionless
    #   google-genai
    #   gradio
    #   gradio-client
    #   grpcio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   starlette
    #   typer
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via
    #   botocore
    #   requests
uvicorn==0.40.0
    # via gradio
validators==0.35.0
    # via frictionless
vertexai==1.71.1
    # via -r requirements.in
yarl==1.22.0
    # via aiohttp
